pyproject.toml
[project]
name = "salecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioboto3>=15.5.0",
    "aiohttp>=3.13.3",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "catboost>=1.2.8",
    "celery[redis]>=5.6.2",
    "fastapi>=0.133.1",
    "flower>=2.0.1",
    "hdbscan>=0.8.41",
    "matplotlib>=3.10.8",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "ruff>=0.14.11",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "sqlalchemy[asyncio]>=2.0.46",
    "statsforecast>=2.0.1",
    "streamlit>=1.54.0",
    "umap-learn>=0.5.9.post2",
    "uvicorn[standard]>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov>=7.0.0",
    "mypy>=1.19.1",
    "kaleido>=1.2.0",
]
jupyter = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "nbformat>=5.10.4",
]
neural = [
    "chronos-forecasting>=2.0",
    "neuralforecast>=3,<4",
]

